+++ src/common/src/price/record.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 pub struct PriceRecord {
     pub currency: CurrencyRecord,
-    pub amount: f32,
+    pub amount: u64,
 }
 
 impl From<Price> for PriceRecord {
